+++ gulpfile.js/tasks/styles_min.js
@@ -1,7 +1,14 @@
 var globals = require('../globals.js')
 
+var src = [
+	globals.getPath( globals.config.paths.dist.styles.all ),
+	'!' + globals.getPath( globals.config.paths.dist.styles.minified.all )
+]
+
+var dest = globals.getPath( globals.config.paths.dist.styles.path )
+
 function stylesMin() {
-	return globals.gulp.src( [globals.config.paths.dist.styles.all, '!'+globals.config.paths.dist.styles.minified.all] )
+	return globals.gulp.src( src )
 		.pipe(
 			globals.$.plumber({
 				errorHandler: globals.consoleError
@@ -13,7 +20,7 @@ function stylesMin() {
 				path.basename += '.min';
 			})
 		)
-		.pipe( globals.gulp.dest( globals.config.paths.dist.styles.path ) );
+		.pipe( globals.gulp.dest( dest ) );
 };
 
 globals.gulp.task('styles:min', stylesMin);